requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
bcrypt
alembic
psycopg2
pydantic-settings 
python-dotenv
PyJWT